+++ src/routes/Auth.js
@@ -1,37 +1,8 @@
 import { authService, firebaseInstance } from "fbase"
-import React, { useState } from "react"
+import React from "react"
+import AuthForm from "components/AuthForm"
 
 const Auth = () => {
-  const [email, setEmail] = useState("")
-  const [password, setPassword] = useState("")
-  const [newAccount, setNewAccount] = useState(true)
-  const [error, setError] = useState("")
-  const onChange = (event) => {
-    const {target: {name, value}} = event
-    
-    if(name === "email") {
-      setEmail(value)
-    } else if(name === "password") {
-      setPassword(value)
-    }
-  }
-  
-  const onSubmit = async (event) => {
-    event.preventDefault()
-    
-    try {
-      if(newAccount) {
-        await authService.createUserWithEmailAndPassword(email, password)
-      } else {
-        await authService.signInWithEmailAndPassword(email, password)
-      }
-    } catch(error) {
-      setError(error.message)
-    }
-  }
-
-  const toggleAccount = () => setNewAccount(prev => !prev)
-
   const onSocialClick = async (event) => {
     const {
       target:{name}
@@ -49,27 +20,7 @@ const Auth = () => {
 
   return (
     <div>
-      <form onSubmit={onSubmit}>
-        <input 
-          name="email" 
-          type="email" 
-          placeholder="Email" 
-          required 
-          value={email}
-          onChange={onChange}
-        />
-        <input 
-          name="password" 
-          type="password" 
-          placeholder="Password" 
-          required 
-          value={password}
-          onChange={onChange}
-        />
-        <input type="submit" value={newAccount ? "Create Account" : "Log In"} />
-        {error}
-      </form>
-      <span onClick={toggleAccount}>{newAccount ? "Log in" : "Create Account "}</span>
+      <AuthForm />
       <div>
         <button onClick={onSocialClick} name="google">Continue with Google</button>
         <button onClick={onSocialClick} name="github">Continue with Github</button>
